ai-engine/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
pymupdf
sentence-transformers
numpy
pydantic
protobuf
python-dotenv
requests